+++ lang/es.json
@@ -1650,5 +1650,8 @@
 	"label.sign_gnosis_safe": "Firmar Gnosis Safe",
 	"label.you_need_to_execute_the_pending_multisig": "Debes ejecutar la transacción Multisig pendiente para completar tu inicio de sesión en Giveth y proceder a esta área",
 	"label.sign_a_message_with_your_safe_signer": "Firma un mensaje con la dirección de tu firma segura (Safe signer) para continuar el proceso de inicio de sesión",
-	"label.this_is_necessary_to_create_projects": "Esto es necesario para poder crear proyectos, gestionar tu perfil o usar GIVpower."
+	"label.this_is_necessary_to_create_projects": "Esto es necesario para poder crear proyectos, gestionar tu perfil o usar GIVpower.",
+	"label.sanctioned_wallet": "¡Dirección Sancionada Detectada!",
+	"label.sanctioned_wallet_message_part1": "Esta dirección ha sido encontrada en la lista de sanciones de EE.UU.",
+	"label.sanctioned_wallet_message_part2": "Lamentablemente, Endaoment no permite que las direcciones en la lista de sanciones de OFAC donen a proyectos entregados por Endaoment. Consulta otro proyecto para donar."
 }
