+++ package.json
@@ -32,12 +32,10 @@
 		"@solana/web3.js": "^1.90.0",
 		"@superfluid-finance/sdk-core": "^0.7.0",
 		"@tanstack/react-query": "^5.45.1",
-		"@types/axios": "^0.14.0",
 		"@vercel/speed-insights": "^1.0.9",
 		"@web3auth/sign-in-with-solana": "^3.0.0",
 		"@web3modal/wagmi": "^5.0.7",
 		"apollo-upload-client": "^18.0.1",
-		"axios": "^1.7.3",
 		"base58": "^2.0.1",
 		"bignumber.js": "^9.1.2",
 		"deepmerge": "^4.3.1",
